feat: improve linux benchmarks by 20% and fix benchmark script

+++ src/fs/buffer.rs
@@ -169,7 +169,7 @@ where
     /// Executes the getdents64 system call using <unistd.h>/direct `libc` syscalls
     #[inline]
     #[cfg(any(target_os = "linux", target_os = "android"))]
-    pub fn getdents(&mut self, fd: &crate::fs::FileDes) -> i64 {
+    pub fn getdents(&mut self, fd: &crate::fs::FileDes) -> isize {
         // SAFETY: we're passing a valid buffer
         unsafe { crate::util::getdents(fd.0, self.as_mut_ptr(), SIZE) }
     }
